+++ packages/workbox/lib/options.js
@@ -3,8 +3,6 @@ const path = require('path')
 const defaults = require('./defaults')
 const { joinUrl, getRouteParams, startCase } = require('@nuxtjs/pwa-utils')
 
-const HMRRegex = '(?!.*(__webpack_hmr|hot-update))'
-
 function getOptions (moduleOptions) {
   const options = Object.assign({}, defaults, moduleOptions, this.options.workbox)
 
@@ -40,23 +38,23 @@ function getOptions (moduleOptions) {
 
   // Cache all _nuxt resources at runtime
   if (!options.assetsURLPattern) {
-    options.assetsURLPattern = joinUrl(options.publicPath, HMRRegex)
+    options.assetsURLPattern = options.publicPath
   }
   if (options.cacheAssets) {
     options.runtimeCaching.push({
       urlPattern: options.assetsURLPattern,
-      handler: 'CacheFirst'
+      handler: this.options.dev ? 'NetworkFirst' : 'CacheFirst'
     })
   }
 
   // Optionally cache other routes for offline
   if (!options.pagesURLPattern) {
-    options.pagesURLPattern = joinUrl(options.routerBase, HMRRegex)
+    options.pagesURLPattern = options.routerBase
   }
   if (options.offline && !options.offlinePage) {
     options.runtimeCaching.push({
       urlPattern: options.pagesURLPattern,
-      handler: 'NetworkFirst'
+      handler: options.offlineStrategy
     })
   }