+++ test/pwa.test.js
@@ -37,8 +37,8 @@ describe('pwa', () => {
   })
 
   test('accessible icons', async () => {
-    const { html } = await nuxt.renderRoute('/icons')
-    expect(html).toContain('/_nuxt/icons/icon_512.9mgd2ZDMcQu.png')
+    const { html } = await nuxt.renderRoute('/')
+    expect(html).toContain('/_nuxt/icons/icon_512.b8f3a1.png')
   })
 
   test('sw.js', async () => {